refactor: Move IriConverter into CoreModule
