+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/gravsearch/transformers/ConstructTransformer.scala
@@ -9,8 +9,7 @@ import zio.*
 
 import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.util.search.*
-import org.knora.webapi.slice.resourceinfo.domain.IriConverter
-
+import org.knora.webapi.slice.ontology.domain.service.IriConverter
 final case class ConstructTransformer(
   sparqlTransformerLive: OntologyInferencer,
   iriConverter: IriConverter,
