+++ webapi/src/main/scala/org/knora/webapi/slice/ontology/domain/service/IriConverter.scala
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.knora.webapi.slice.resourceinfo.domain
+package org.knora.webapi.slice.ontology.domain.service
 
 import zio.IO
 import zio.Task
@@ -15,6 +15,7 @@ import org.knora.webapi.messages.SmartIri
 import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.slice.common.KnoraIris.PropertyIri
 import org.knora.webapi.slice.common.KnoraIris.ResourceClassIri
+import org.knora.webapi.slice.resourceinfo.domain.InternalIri
 
 final case class IriConverter(sf: StringFormatter) {
   def asSmartIri(iri: String): Task[SmartIri]              = ZIO.attempt(sf.toSmartIri(iri, requireInternal = false))
@@ -44,5 +45,5 @@ final case class IriConverter(sf: StringFormatter) {
 }
 
 object IriConverter {
-  val layer: ZLayer[StringFormatter, Nothing, IriConverter] = ZLayer.derive[IriConverter]
+  val layer = ZLayer.derive[IriConverter]
 }
