+++ webapi/src/test/scala/org/knora/webapi/slice/admin/domain/service/LegalInfoServiceSpec.scala
@@ -19,8 +19,8 @@ import org.knora.webapi.slice.admin.domain.model.License
 import org.knora.webapi.slice.admin.domain.model.LicenseIri
 import org.knora.webapi.slice.admin.domain.repo.KnoraProjectRepoInMemory
 import org.knora.webapi.slice.admin.repo.LicenseRepo
+import org.knora.webapi.slice.ontology.domain.service.IriConverter
 import org.knora.webapi.slice.ontology.repo.service.OntologyRepoInMemory
-import org.knora.webapi.slice.resourceinfo.domain.IriConverter
 import org.knora.webapi.store.triplestore.api.TriplestoreServiceInMemory
 
 object LegalInfoServiceSpec extends ZIOSpecDefault {
