+++ webapi/src/test/scala/org/knora/webapi/slice/ontology/repo/service/OntologyRepoLiveSpec.scala
@@ -16,9 +16,9 @@ import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.slice.ontology.domain.OntologyCacheDataBuilder
 import org.knora.webapi.slice.ontology.domain.ReadClassInfoV2Builder
 import org.knora.webapi.slice.ontology.domain.ReadOntologyV2Builder
+import org.knora.webapi.slice.ontology.domain.service.IriConverter
 import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
 import org.knora.webapi.slice.resourceinfo.domain.InternalIri
-import org.knora.webapi.slice.resourceinfo.domain.IriConverter
 import org.knora.webapi.slice.resourceinfo.domain.IriTestConstants.Anything
 import org.knora.webapi.slice.resourceinfo.domain.IriTestConstants.Biblio
